Introduce `Microsoft.NAV` requests as a concept
According to the Dynamics API docs, there is a standard interface for
non-CRUD actions on an object called `Microsoft.NAV`. This commit
introduces this as a concept, and implements it for salesInvoices, where
the method `Microsoft.NAV.post` is used to "post" an invoice within
Dynamics (which means converts it from a draft to an open invoice, ready
for payment).